requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
gunicorn==23.0.0
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
redis==5.1.1
celery==5.4.0
pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
weasyprint==62.3
groq==0.11.0
pydicom==3.0.1
python-docx==1.1.2
openpyxl==3.1.1
reportlab==4.2.5
alembic==1.13.3
pytest==8.3.3
